sorta working

+++ model/unit.go
@@ -1,7 +1,7 @@
 package model
 
 type Unit struct {
-	PieceId     int      `json:"piece_id",omitempty`
+	PieceId     int64    `json:"piece_id",omitempty`
 	UnitType    UnitType `json:"unit_type"`
 	Owner       Country  `json:owner"`
 	WillRetreat bool     `json:"will_retreat,omitempty"`
